NMEA/EngineState: convert revs_per_sec to float
Having it as integer means we can only present multiples of 60rpm, and
thus we have to discard precision from the IGC fix.

+++ android/src/NativeSensorListener.java
@@ -39,7 +39,7 @@ final class NativeSensorListener implements SensorListener {
                                      boolean has_egt_temp,
                                      int egt_temp,
                                      boolean has_revs_per_sec,
-                                     int revs_per_sec);
+                                     float revs_per_sec);
 
   @Override
   public native void onAccelerationSensor(float ddx, float ddy, float ddz);
