+++ android/src/SensorListener.java
@@ -47,7 +47,7 @@ public interface SensorListener {
                        boolean has_egt_temp,
                        int egt_temp,
                        boolean has_revs_per_sec,
-                       int revs_per_sec);
+                       float revs_per_sec);
 
   void onVoltageValues(int temp_adc, int voltage_index, int volt_adc);
